Cadastro agora não aceita CPF e EMAIL já cadastrados

+++ acaoCadastro.php
@@ -4,21 +4,34 @@ echo '<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min
 
 
 function inserir($conexao, $nome, $sobrenome, $data_nascimento, $sexo, $cpf, $cep, $telefone, $endereco, $num_endereco, $complemento, $cidade, $bairro, $estado, $email, $senha) {
+
+	$script = "SELECT * FROM usuarios WHERE cpf = $cpf";
+	$consulta = $conexao->query($script);
+	if ($linha = $consulta->fetch_array(MYSQLI_ASSOC)) {
+		echo "<script> $.confirm({type: 'red', title: 'Cadastro', content: 'CPF já cadastrado.', buttons: { Ok: { btnClass: 'btn-red', action: function () {"; echo 'window.location="cadastro.php"'; echo " }}}});</script>";
+	} else {
+		$script = 'SELECT * FROM usuarios WHERE email = "' . $email . '"';
+		$consulta = $conexao->query($script);
+		if ($linha = $consulta->fetch_array(MYSQLI_ASSOC)) {
+			echo "<script> $.confirm({type: 'red', title: 'Cadastro', content: 'Email já cadastrado.', buttons: { Ok: { btnClass: 'btn-red', action: function () {"; echo 'window.location="cadastro.php"'; echo " }}}});</script>";
+		} else {
 	
-	$script = 'INSERT INTO usuarios VALUES (DEFAULT, "' . 0 . '", "' . $nome . '", "' . $sobrenome . '","' . $data_nascimento . '","' . $sexo . '", "' . $cpf . '", "' . $cep . '", "' . $telefone . '", "' . $endereco . '", "' . $num_endereco . '", "' . $complemento . '", "' . $cidade . '", "' . $bairro . '", "' . $estado . '", "' . $email . '", "' . $senha . '")';
+			$script = 'INSERT INTO usuarios VALUES (DEFAULT, "' . 0 . '", "' . $nome . '", "' . $sobrenome . '","' . $data_nascimento . '","' . $sexo . '", "' . $cpf . '", "' . $cep . '", "' . $telefone . '", "' . $endereco . '", "' . $num_endereco . '", "' . $complemento . '", "' . $cidade . '", "' . $bairro . '", "' . $estado . '", "' . $email . '", "' . $senha . '")';
 
-	$insere = $conexao->query($script);
+			$insere = $conexao->query($script);
 
-	if(!$insere){
-		echo "<br>Inserção incorreta! Deu erro!<br>";
-		echo $conexao->error;
-		echo "<br><br>";
-		echo $script;
-		echo "<br>";
-		echo "<script> $.confirm({type: 'red', title: 'Cadastro', content: 'Cadastro incorreto. Favor verificar dados.', buttons: { Ok: { btnClass: 'btn-red', action: function () {"; echo 'window.location="cadastro.php"'; echo " }}}});</script>";
-	}else{
-		echo "<br>Inserção Realizada corretamente!";
-		echo "<script> $.confirm({type: 'red', title: 'Cadastro', content: 'Cadastro Concluído! Faça login para continuar.', buttons: { Ok: { btnClass: 'btn-red', action: function () {"; echo 'window.location="login.php"'; echo " }}}});</script>";
+			if(!$insere){
+				echo "<br>Inserção incorreta! Deu erro!<br>";
+				echo $conexao->error;
+				echo "<br><br>";
+				echo $script;
+				echo "<br>";
+				echo "<script> $.confirm({type: 'red', title: 'Cadastro', content: 'Cadastro incorreto. Favor verificar dados.', buttons: { Ok: { btnClass: 'btn-red', action: function () {"; echo 'window.location="cadastro.php"'; echo " }}}});</script>";
+			}else{
+				echo "<br>Inserção Realizada corretamente!";
+				echo "<script> $.confirm({type: 'red', title: 'Cadastro', content: 'Cadastro Concluído! Faça login para continuar.', buttons: { Ok: { btnClass: 'btn-red', action: function () {"; echo 'window.location="login.php"'; echo " }}}});</script>";
+			}
+		}
 	}
 }
